+++ Publicizer.Tests/Data/OtherTypes.cs
@@ -1,4 +1,5 @@
-﻿namespace NamespaceForOtherTypes;
+﻿// NOTE: namespace intentionally differs from folder structure to test the code generation
+namespace NamespaceForOtherTypes;
 
 public struct OtherType
 {
